fix: calculate selection area with zoom level too properly calculate selection area bounds but with zoom level taken into account.
note: the zoom factor is defined as `window.devicePixelRatio`, which may
be problematic for high-DPI displays like Retina displays. I do not have
the ability to test this.

note 2: there was another quantity, `window.visualViewport.scale` that
is relevant for mobile users "pinching" to zoom the screen. this will
have to be considered in order to implement mobile support.